+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 1. [](#bugfix)
     * Fix foreach-loop in Transport->setStyle()
+    * Fix textsizing-class application
 
 # v1.2.0
 ## 14-01-2019